+++ packages/components/CHANGELOG.md
@@ -16,7 +16,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 
 ### Added
 
-- Vue: Custom form components support `v-model` with argument `value|checked`. Can be used like this:
+- Angular: form components implement ControlValueAccessor interface to support Angular forms out of the box
+  ([#4048](https://github.com/porsche-design-system/porsche-design-system/pull/4048))
+- Vue: form components support `v-model` with argument `value|checked`. Can be used like this:
   `<PInputText v-model:value="foo" />`
   ([#4066](https://github.com/porsche-design-system/porsche-design-system/pull/4066))
 - `Pin Code`, `Segmented Control`:
@@ -38,6 +40,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 
 ### Fixed
 
+- `Textarea`, `Input Password`: `input` event `target` doesn't have updated `value`
+  ([#4048](https://github.com/porsche-design-system/porsche-design-system/pull/4048))
 - `Select Wrapper`: disabled options in native dropdown not displayed correctly
   ([#4046](https://github.com/porsche-design-system/porsche-design-system/pull/4046))
 - `Tailwind`: change `--color-base` to `--color-canvas` to avoid conflicts with `text-base`